[feat] Add simple code coverage tool
This commit introduces a basic API for the CoverageChecker object which
can be used to measure code coverage manually. The API is very simple,
as it contains two methods: an initialization method that is idempotent
and must be called in the start of the function that is being tested,
and a method that is used to indicate that a branch with a certain index
has been reached.

The commit also includes an example of usage through the CoverageExample
object. As the CoverageChecker is an object, it is essentially a static
class in Java terms and can be imported and used directly without
instantiation.

[Issue: #1]