[feat] Add test of default case in Config.collect
The method Config.collect was previously missing test coverage for the
case where an invoked method did not return a value that matched any
case, thus leading to an empty default case. In this situation, the
method is not supposed to recursively re-invoke the collect method. This
is tested by having a method that populates a defaulted value and
returns None, leading to a situation where if the buffer is recursively
checked again after the method has been invoked, the test will fail.

[Issue: #30]