[feat] Add test for case Promise in setInterruptHandler
The method Promise.setInterruptHandler is supposed to propagate to any
linked promises, by setting the same interrupt handler to those
promises. This behaviour was previously untested, but through this
commit a test has been created to ensure this functionality.

The test features heavy use of reflection in order to modify the private
field state in the Promise instance, which is necessary as there does
not seem to be any available API for changing the value of that field to
an instance of Promise.

[Issue: #32]