[feat] Add test for Some case in Config.collect (#36)
The method Config.collect was previously missing test coverage for the
case where an invoked method returns Some(Config[_]). In this case, the
program is supposed to unwrap the monad and recursively search for
missing values in the inner configuration. This has been tested by
adding a simple class method that populates a missing value if executed
and then returns itself, wrapped in an Option monad. If the program
recurses after executing the method, it will find the populated missing
value. If it does not, the value will still be missing.

[Issue: #29]